Finished new table writer, implemented batch load flow in BigQuerySinkTask
Finished the new table writer and altered it to work similarly to the normal table writer.
Introduced a GCS writer.
Added a configuration for the name of the bucket where GCS blobs should be uploaded.
Altered the flow of BigQuerySinkTask to batch process records in topics configured to be batch loaded